+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/CustomAriaLabel.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Select from '../../../Select';
-import SelectField from '../../../SelectField';
+import Select from 'terra-form-select/lib/Select';
+import SelectField from 'terra-form-select/lib/SelectField';
 import styles from './common/Select.test.module.scss';
 
 const cx = classNames.bind(styles);