+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/IsTouchAccessibleTagSelect.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import TagSelect from '../../../TagSelect';
+import TagSelect from 'terra-form-select/lib/TagSelect';
 import styles from './common/Select.test.module.scss';
 
 const cx = classNames.bind(styles);