+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/SelectInsideIframe.test.jsx
@@ -6,7 +6,7 @@ const cx = classNames.bind(styles);
 
 const SelectInsideIframe = () => (
   <div id="frame-container" className={cx('content-wrapper')}>
-    <iframe id="frmSelectFrame" title="form-selectFrame" src="/#/raw/tests/terra-form-select/form-select/many-options" height="650" width="355" />
+    <iframe id="frmSelectFrame" title="form-selectFrame" src="/#/raw/tests/cerner-terra-core-docs/form-select/many-options" height="650" width="355" />
   </div>
 );