+++ packages/terra-core-docs/src/terra-dev-site/test/form-select/SingleSelectFieldMaxheight.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
+import SingleSelectField from 'terra-form-select/lib/SingleSelectField';
 import styles from './common/Select.test.module.scss';
-import SingleSelectField from '../../../SingleSelectField';
 
 const cx = classNames.bind(styles);