+++ packages/terra-form-select/tests/wdio/_frame-spec.js
@@ -1,6 +1,6 @@
 Terra.describeViewports('Frame', ['tiny'], () => {
   it('should have a themeable disabled icon', () => {
-    browser.url('/raw/tests/terra-form-select/form-select/disabled');
+    browser.url('/raw/tests/cerner-terra-core-docs/form-select/disabled');
 
     Terra.validates.element('themeable disabled icon');
   });