+++ packages/terra-core-docs/CHANGELOG.md
@@ -9,6 +9,7 @@
   * Migrate docs from terra-form-field.
   * Migrate docs from terra-divider.
   * Migrate docs from terra-form-fieldset.
+  * Migrate docs from terra-form-select.
 
 ## 1.12.0 - (November 2, 2021)