+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Changed
+  * Migrate doc to terra-core-docs.
+
+* Changed
   * Updated dependencies to not be hard-coded.
 
 ## 6.35.0 - (September 21, 2021)